remove game class from main.rb
